// src/modules/store-catalog/facade/store-catalog.facade.interface.ts
/* Qualquer parte da aplicação que quiser se comunicar com esse
módulo utilizará a interface que está definida nessa facade. */

export interface FindStoreCatalogFacadeInputDto {
    id: string;
}

export interface FindStoreCatalogFacadeOutputDto {
    id: string;
    name: string;
    description: string;
    salesPrice: number;
}

export interface FindAllStoreCatalogFacadeOutputDto {
    products: {
        id: string;
        name: string;
        description: string;
        salesPrice: number;
    }[]
}

export default interface StoreCatalogFacadeInterface {

    find(id: FindStoreCatalogFacadeInputDto): Promise<FindStoreCatalogFacadeOutputDto>

    findAll(): Promise<FindAllStoreCatalogFacadeOutputDto>;
}